feat: implement rune efficiency and contribution calculations

+++ app/src/main/java/com/example/swrunevault/models/SubStatRange.kt
@@ -3,8 +3,8 @@ package com.example.swrunevault.models
 // Rango posible de una subestadística para una cantidad determinada de estrellas.
 data class SubStatRange(
     // Valor mínimo posible.
-    val minValue: Double,
+    val minValue: Int,
 
     // Valor máximo posible.
-    val maxValue: Double
+    val maxValue: Int
 )
